Fix and restore grpc_pass call for the GRPC API
This was removed by accident; it's still needed for the API but we
changed the location matcher to ensure that we don't swallow actual
HTTP API calls since both the API and the main app use the same
nginx configuration. We could be smarter by only using the grpc_pass
call on the chirpstack proxy, not the API proxy, but our current
docker-compose setup doesn't really allow for that easily.